Fix narrowing conversions and update camera info manager API
Parameter values arrive as 64-bit integers, so retrieve them as int64_t
and narrow them explicitly. V4L2 controls and image dimensions are
range-checked before narrowing, and out-of-range requests are rejected
with a warning instead of silently wrapping.

Also pass the RMW QoS profile to CameraInfoManager, matching the
constructor signature available in this distribution.

+++ include/v4l2_camera/parameters.hpp
@@ -79,7 +79,10 @@ public:
 
   std::string getOutputEncoding() const { return getValue<std::string>("output_encoding"); }
 
-  int getCvtColorNumThreads() const { return getValue<int>("cvt_color_num_threads"); }
+  int getCvtColorNumThreads() const
+  {
+    return static_cast<int>(getValue<int64_t>("cvt_color_num_threads"));
+  }
 
   std::string getPixelFormat() const { return getValue<std::string>("pixel_format"); }
   std::vector<int64_t> getImageSize() const { return getValue<std::vector<int64_t>>("image_size"); }
